CRUD App complete

+++ django_crud_app/student/crud/views.py
@@ -1,9 +1,5 @@
-import re
-from tkinter import N
-from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from .models import Student
-# Create your views here.
 
 def home(request):
     students = Student.objects.all()
@@ -11,16 +7,25 @@ def home(request):
 
     return render(request, "home.html", context)
 
-def deletestudent(id):
-    student = Student.objects.get(id=id)
-    student.delete()
+def deletestudent(request, id):
+    if request.method == "POST":
+        student = Student.objects.get(id=id)
+        student.delete()
     return redirect("home")
 
 def createstudent(request):
     if request.method == "POST":
         name = request.POST.get("name")
         fee = request.POST.get("fee")
+        Student.objects.create(name=name,fee=fee)
+    return redirect("home")
 
-        student = Student(name=name,fee=fee)
+def updatestudent(request,id):
+    if request.method == "POST":
+        student = Student.objects.get(id=id)        
+        name = request.POST.get("name")
+        fee = request.POST.get("fee")
+        student.name = name
+        student.fee = fee
         student.save()
-    return redirect("home") 
+    return redirect("home")
